fix(family/09): route.ts の schema export を削除しビルドエラーを解消
Next.js 14 App Router の route.ts は HTTP method handler (GET/POST 等) と
予約済 config 以外の export を許可しないため、Zod schema / type を別 lib
ファイルに切り出して import 形に変更。Vercel デプロイがブロックされていた
"HandsonTourCompleteResponseSchema is not a valid Route export field" を解消。

変更:
- src/lib/handson-tour/schemas.ts 新規 (Status/Complete/Skip の Request/Response
  schema と type を集約 export)
- src/app/api/handson-tour/skip/route.ts: schema/type 定義を削除し schemas.ts
  から import に
- src/app/api/handson-tour/complete/route.ts: 同上

設計書 family/02 §15 では schema を route.ts inline 定義としていたが、
Next.js 制約により lib 切り出しが正解。後続 PR で family/02 §15 と family/09 §09 に
schemas.ts 経路の運用ルール追記予定。
